BaseTest class added ListTest class added Some minor changes made
