+++ src/main/java/com/epam/tc/hw9/Constants.java
@@ -1,5 +1,7 @@
 package com.epam.tc.hw9;
 
+import com.epam.tc.hw9.utils.ApiProperties;
+
 public class Constants {
     public static final String KEY = "key";
     public static final String TOKEN = "token";
