+++ src/main/java/com/epam/tc/hw9/beans/Prefs.java
@@ -2,7 +2,6 @@ package com.epam.tc.hw9.beans;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
-import java.util.Objects;
 
 public class Prefs {
     @SerializedName("background")
