+++ src/test/resources/test.properties
@@ -1,7 +1,7 @@
 driver=chrome
 
-key=a2d5821da3546b13fdaf5872cb4ea8a7
-token=af3e31838be892bd382e26827f63dcb2cd0b84a895b12b8977f76762c9aae1d5
+key=
+token=
 root_path=https://api.trello.com/1
 endpoint=/boards/
 lists=/lists
